Updated SkillCollection to remove circular refs
Read methods of SkillCollection have been moved to ReadOnlySkillCollection
ReadOnlySkillCollection has access to underlying skill collections dict
