+++ python/semantic_kernel/skill_definition/skill_collection.py
@@ -1,11 +1,17 @@
 # Copyright (c) Microsoft. All rights reserved.
 
 from logging import Logger
-from typing import TYPE_CHECKING, Dict, Generic, Literal, Optional, Tuple, TypeVar
+from typing import (
+    TYPE_CHECKING,
+    Dict,
+    Generic,
+    Optional,
+    TypeVar,
+    Union,
+)
 
 import pydantic as pdt
 
-from semantic_kernel.kernel_exception import KernelException
 from semantic_kernel.sk_pydantic import SKGenericModel
 from semantic_kernel.skill_definition.functions_view import FunctionsView
 from semantic_kernel.skill_definition.read_only_skill_collection import (
@@ -13,35 +19,36 @@ from semantic_kernel.skill_definition.read_only_skill_collection import (
 )
 from semantic_kernel.skill_definition.skill_collection_base import SkillCollectionBase
 from semantic_kernel.utils.null_logger import NullLogger
-from semantic_kernel.utils.static_property import static_property
 
 if TYPE_CHECKING:
     from semantic_kernel.orchestration.sk_function_base import SKFunctionBase
-    from semantic_kernel.skill_definition.read_only_skill_collection_base import (
-        ReadOnlySkillCollectionBase,
-    )
 
 
 SKFunctionT = TypeVar("SKFunctionT", bound="SKFunctionBase")
 
 
 class SkillCollection(SKGenericModel, SkillCollectionBase, Generic[SKFunctionT]):
-    skill_collection: Dict[str, Dict[str, "SKFunctionBase"]]
-    _read_only_skill_collection: "ReadOnlySkillCollectionBase" = pdt.PrivateAttr()
+    skill_collection: Dict[str, Dict[str, SKFunctionT]]
+    _read_only_skill_collection: ReadOnlySkillCollection = pdt.PrivateAttr(default=None)
     _log: Logger = pdt.PrivateAttr()
-
-    @property
-    def read_only_skill_collection(self) -> "ReadOnlySkillCollectionBase":
-        return self._read_only_skill_collection
 
     def __init__(
         self,
         log: Optional[Logger] = None,
-        skill_collection: Optional[Dict[str, Dict[str, SKFunctionT]]] = None,
+        skill_collection: Union[Dict[str, Dict[str, SKFunctionT]], None] = None,
     ) -> None:
-        super().__init__(skill_collection=skill_collection or {})
+        if skill_collection is None:
+            skill_collection = {}
+        super().__init__(skill_collection=skill_collection)
         self._log = log if log is not None else NullLogger()
-        self._read_only_skill_collection = ReadOnlySkillCollection(self)
+
+    @property
+    def read_only_skill_collection(self) -> ReadOnlySkillCollection:
+        if self._read_only_skill_collection is None:
+            self._read_only_skill_collection = ReadOnlySkillCollection(
+                self.skill_collection
+            )
+        return self._read_only_skill_collection
 
     def add_semantic_function(self, function: "SKFunctionBase") -> None:
         if function is None:
@@ -60,7 +67,9 @@ class SkillCollection(SKGenericModel, SkillCollectionBase, Generic[SKFunctionT])
             raise ValueError("The function provided cannot be `None`")
 
         s_name, f_name = function.skill_name, function.name
-        s_name, f_name = self._normalize_names(s_name, f_name, True)
+        s_name, f_name = self.read_only_skill_collection._normalize_names(
+            s_name, f_name, True
+        )
 
         if s_name not in self.skill_collection:
             self.skill_collection[s_name] = {}
@@ -68,96 +77,44 @@ class SkillCollection(SKGenericModel, SkillCollectionBase, Generic[SKFunctionT])
         self.skill_collection[s_name][f_name] = function
 
     def has_function(self, skill_name: Optional[str], function_name: str) -> bool:
-        s_name, f_name = self._normalize_names(skill_name, function_name, True)
-        if s_name not in self.skill_collection:
-            return False
-        return f_name in self.skill_collection[s_name]
+        return self.read_only_skill_collection.has_function(skill_name, function_name)
 
-    def has_semantic_function(self, skill_name: str, function_name: str) -> bool:
-        s_name, f_name = self._normalize_names(skill_name, function_name)
-        if s_name not in self.skill_collection:
-            return False
-        if f_name not in self.skill_collection[s_name]:
-            return False
-        return self.skill_collection[s_name][f_name].is_semantic
+    def has_semantic_function(
+        self, skill_name: Optional[str], function_name: str
+    ) -> bool:
+        return self.read_only_skill_collection.has_semantic_function(
+            skill_name, function_name
+        )
 
-    def has_native_function(self, skill_name: str, function_name: str) -> bool:
-        s_name, f_name = self._normalize_names(skill_name, function_name, True)
-        if s_name not in self.skill_collection:
-            return False
-        if f_name not in self.skill_collection[s_name]:
-            return False
-        return self.skill_collection[s_name][f_name].is_native
+    def has_native_function(
+        self, skill_name: Optional[str], function_name: str
+    ) -> bool:
+        return self.read_only_skill_collection.has_native_function(
+            skill_name, function_name
+        )
 
     def get_semantic_function(
-        self, skill_name: str, function_name: str
+        self, skill_name: Optional[str], function_name: str
     ) -> "SKFunctionBase":
-        s_name, f_name = self._normalize_names(skill_name, function_name)
-        if self.has_semantic_function(s_name, f_name):
-            return self.skill_collection[s_name][f_name]
-
-        self._log.error(f"Function not available: {s_name}.{f_name}")
-        raise KernelException(
-            KernelException.ErrorCodes.FunctionNotAvailable,
-            f"Function not available: {s_name}.{f_name}",
+        return self.read_only_skill_collection.get_semantic_function(
+            skill_name, function_name
         )
 
     def get_native_function(
-        self, skill_name: str, function_name: str
+        self, skill_name: Optional[str], function_name: str
     ) -> "SKFunctionBase":
-        s_name, f_name = self._normalize_names(skill_name, function_name, True)
-        if self.has_native_function(s_name, f_name):
-            return self.skill_collection[s_name][f_name]
-
-        self._log.error(f"Function not available: {s_name}.{f_name}")
-        raise KernelException(
-            KernelException.ErrorCodes.FunctionNotAvailable,
-            f"Function not available: {s_name}.{f_name}",
+        return self.read_only_skill_collection.get_native_function(
+            skill_name, function_name
         )
 
     def get_functions_view(
         self, include_semantic: bool = True, include_native: bool = True
     ) -> FunctionsView:
-        result = FunctionsView()
-
-        for skill in self.skill_collection.values():
-            for function in skill.values():
-                if include_semantic and function.is_semantic:
-                    result.add_function(function.describe())
-                elif include_native and function.is_native:
-                    result.add_function(function.describe())
-
-        return result
+        return self.read_only_skill_collection.get_functions_view(
+            include_semantic, include_native
+        )
 
     def get_function(
         self, skill_name: Optional[str], function_name: str
     ) -> "SKFunctionBase":
-        s_name, f_name = self._normalize_names(skill_name, function_name, True)
-        if self.has_function(s_name, f_name):
-            return self.skill_collection[s_name][f_name]
-
-        self._log.error(f"Function not available: {s_name}.{f_name}")
-        raise KernelException(
-            KernelException.ErrorCodes.FunctionNotAvailable,
-            f"Function not available: {s_name}.{f_name}",
-        )
-
-    def _normalize_names(
-        self,
-        skill_name: Optional[str],
-        function_name: str,
-        allow_substitution: bool = False,
-    ) -> Tuple[str, str]:
-        s_name, f_name = skill_name, function_name
-        if s_name is None and allow_substitution:
-            s_name = self.GLOBAL_SKILL
-
-        if s_name is None:
-            raise ValueError("The skill name provided cannot be `None`")
-
-        s_name, f_name = s_name.lower(), f_name.lower()
-        return s_name, f_name
-
-    @static_property
-    def GLOBAL_SKILL() -> Literal["_GLOBAL_FUNCTIONS_"]:
-        return "_GLOBAL_FUNCTIONS_"
+        return self.read_only_skill_collection.get_function(skill_name, function_name)
